added text animation
