+++ src/pages/Home/textAnimation.scss
@@ -1,0 +1,50 @@
+// for react text animations
+.foo {
+    width: 200px;
+    
+    .container {
+      display: inline-block;
+      position: relative;
+      padding: 2px 4px;
+      transition: all 200ms cubic-bezier(0.2, 1, 0.2, 1);
+  
+      &:after {
+        background-color: rgb(52, 93, 160);
+        content: "";
+        position: absolute;
+        top: -2px;
+        right: 100%;
+        bottom: 2px;
+        left: 0;
+        transition: all cubic-bezier(0.2, 1, 0.2, 1) 500ms 250ms;
+      }
+    }
+  
+    &.foo-out {
+      .container:after {
+        right: 0;
+        transition: all cubic-bezier(0.5, -0.5, 0.25, 1.5) 500ms;
+      }
+    }
+  }
+  
+  .text-width-wrapper {
+    display: inline-block;
+  }
+  
+  @keyframes show {
+    from {
+      opacity: 0;
+    }
+    to {
+      opacity: 1;
+    }
+  }
+  
+  .hidden-text {
+    position: absolute;
+    top: -999px;
+    left: -999px;
+  }
+  
+
